refactor: support 64-bit addresses
- Change type of various variables and function parameters from `usize` to `u64` to support 64-bit addresses.

+++ src/memlayout.zig
@@ -28,32 +28,32 @@ pub const virtio0_irq = 1;
 
 // core local interruptor (CLINT), which contains the timer.
 pub const clint = 0x2000000;
-pub inline fn clintMtimecmp(hartid: usize) *usize {
+pub inline fn clintMtimecmp(hartid: u64) *u64 {
     return @ptrFromInt(clint + 0x4000 + 8 * hartid);
 }
 // cycles since boot.
-pub const clint_mtime: *usize = @ptrFromInt(clint + 0xBFF8);
+pub const clint_mtime: *u64 = @ptrFromInt(clint + 0xBFF8);
 
 // qemu puts platform-level interrupt controller (PLIC) here.
 pub const plic = 0x0c000000;
 pub const plic_priority = plic + 0x0;
 pub const plic_pending = plic + 0x1000;
-pub inline fn plicMEnable(hart: usize) usize {
+pub inline fn plicMEnable(hart: u64) u64 {
     return plic + 0x2000 + hart * 0x100;
 }
-pub inline fn plicSEnable(hart: usize) usize {
+pub inline fn plicSEnable(hart: u64) u64 {
     return plic + 0x2080 + hart * 0x100;
 }
-pub inline fn plicMPriority(hart: usize) usize {
+pub inline fn plicMPriority(hart: u64) u64 {
     return plic + 0x200000 + hart * 0x2000;
 }
-pub inline fn plicSPriority(hart: usize) usize {
+pub inline fn plicSPriority(hart: u64) u64 {
     return plic + 0x201000 + hart * 0x2000;
 }
-pub inline fn plicMClaim(hart: usize) usize {
+pub inline fn plicMClaim(hart: u64) u64 {
     return plic + 0x200004 + hart * 0x2000;
 }
-pub inline fn plicSClaim(hart: usize) usize {
+pub inline fn plicSClaim(hart: u64) u64 {
     return plic + 0x201004 + hart * 0x2000;
 }
 
@@ -65,11 +65,11 @@ pub const phy_stop = kernel_base + 128 * 1024 * 1024;
 
 // map the trampoline page to the highest address,
 // in both user and kernel space.
-pub const trampoline: usize = riscv.maxva - @as(usize, riscv.pg_size);
+pub const trampoline: u64 = riscv.maxva - @as(u64, riscv.pg_size);
 
 ///map kernel stacks beneath the trampoline,
 ///each surrounded by invalid guard pages.
-pub inline fn kStack(p: usize) usize {
+pub inline fn kStack(p: u64) u64 {
     return trampoline - (p + 1) * 2 * riscv.pg_size;
 }
 
@@ -82,4 +82,4 @@ pub inline fn kStack(p: usize) usize {
 //   ...
 //   TRAPFRAME (p->trapframe, used by the trampoline)
 //   TRAMPOLINE (the same page as in the kernel)
-pub const trap_frame: usize = trampoline - riscv.pg_size;
+pub const trap_frame: u64 = trampoline - riscv.pg_size;
